Changes: - .gitmodules: googletest now added as git submodule - googletest: googletest now added as git submodule - gtest.pro: googletest now point to local submodule, source files and qt libs added to test project - tst_collisions.h: debouncing test case implemented - monsterchase.cpp: bounding box based offset moved to player.setEnergyGaugePos() - player.cpp/h: states and views moved to separated files (project refactoring) - playersm.cc/h: player states to separated files (project refactoring) - playerqtviews.cpp/h: views moved to separated files (project refactoring) - src.pro: project refactoring
