+++ gtest/gtest.pro
@@ -1,16 +1,19 @@
-GOOGLETEST_DIR = /home/andrea/Playgrounds/CPlusPlus_playground/googletest
+GOOGLETEST_DIR = ../googletest
 
 include(gtest_dependency.pri)
 
 TEMPLATE = app
-CONFIG += console c++11
-#CONFIG += c++11
+#CONFIG += console c++11
+CONFIG += c++11
 CONFIG -= app_bundle
 CONFIG += thread
-CONFIG -= qt
+#CONFIG -= qt
+QT += widgets
+CONFIG += qt
 
 HEADERS += \
         tst_collisions.h
 
 SOURCES += \
-        main.cpp
+        main.cpp \
+        ../src/playersm.cc
